server/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
python-docx
groq
python-dotenv
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-multipart
